telebot3/requirements.txt
python-dotenv
pyodbc
paramiko
pillow
pandas
python-telegram-bot
sqlalchemy
pymssql